applying some changes in mongo_client.py
