+++ db/mongo_client.py
@@ -68,10 +68,7 @@ class Connection:
         
     def update_by_condition(self, condition, new_data):
         try:
-            result = self.collection.update_one(condition, {"$set": new_data})
-            if result.matched_count == 0:
-                return result
-            return result
+            return self.collection.update_one(condition, {"$set": new_data})
         except Exception as e:
             logging.exception(e)
             raise e
@@ -100,16 +97,4 @@ class Connection:
         except Exception as e:
             logging.error(f"Database error in find_by_email: {str(e)}", exc_info=True)
             raise
-
-    def update_password(self, email, new_password):
-        try:
-            result = self.collection.update_one(
-                {"email": email},
-                {"$set": {"password": new_password}}
-            )
-            if result.matched_count == 0:
-                raise Exception("No user found with the provided email.")
-        except Exception as e:
-            logging.error(f"Error updating password: {str(e)}", exc_info=True)
-            raise Exception('Error updating password')
